+++ src/offers/types/offer-response.type.ts
@@ -1,5 +1,6 @@
-import { Offer } from '@prisma/client';
+import { Offer, VoteType } from '@prisma/client';
 
 export type OfferResponse = Offer & {
   createdByUsername: string;
+  userVote: VoteType | null;
 };